refactor-ir-cake: move classes around in sigma.compiler.ir
