+++ sc/shared/src/main/scala/sigma/compiler/ir/Modules.scala
@@ -1,5 +1,7 @@
 package sigma.compiler.ir
 
+import sigma.compiler.ir.meta.ModuleInfo
+
 trait Modules extends Base { self: IRContext =>
 
   /** Whether staged modules should be registered when cake is constructed and initialized. */
