+++ sc/shared/src/main/scala/sigma/compiler/ir/meta/SSymName.scala
@@ -1,8 +1,6 @@
-package scalan.meta
+package sigma.compiler.ir.meta
 
 import sigma.util.StringUtil.StringUtilExtensions
-
-case class ImportItem(packageName: String, importedNames: List[String])
 
 case class SSymName(packageName: String, name: String) {
   import SSymName._
@@ -10,8 +8,6 @@ case class SSymName(packageName: String, name: String) {
 }
 
 object SSymName {
-  /** Wildcard character used to signify importing all names from namespace */
-  val ImportAllWildcard = "*"
   def fullNameString(packageName: String, name: String): String =
     if (packageName.isNullOrEmpty) name else s"$packageName.$name"
 }
