+++ sc/shared/src/main/scala/special/collection/impl/CollsImpl.scala
@@ -7,7 +7,8 @@ import sigma.compiler.ir.IRContext
 import scala.collection.compat.immutable.ArraySeq
 
 package impl {
-  import sigma.compiler.ir.{Base, GraphIRReflection, ModuleInfo, IRContext}
+  import sigma.compiler.ir.meta.ModuleInfo
+  import sigma.compiler.ir.{Base, GraphIRReflection, IRContext}
   import sigma.data.{Nullable, RType}
   import sigma.reflection.{RClass, RMethod}
