+++ sc/shared/src/main/scala/wrappers/scalan/impl/WRTypesImpl.scala
@@ -8,7 +8,8 @@ import special.wrappers.RTypeWrapSpec
 import scala.collection.compat.immutable.ArraySeq
 
 package impl {
-  import sigma.compiler.ir.{Base, GraphIRReflection, ModuleInfo, IRContext}
+  import sigma.compiler.ir.meta.ModuleInfo
+  import sigma.compiler.ir.{Base, GraphIRReflection, IRContext}
   import sigma.reflection.{RClass, RMethod}
 
   // Abs -----------------------------------
