+++ sc/shared/src/main/scala/wrappers/special/impl/WSpecialPredefsImpl.scala
@@ -5,7 +5,9 @@ import scalan._
 import sigma.compiler.ir.IRContext
 
 package impl {
-  import sigma.compiler.ir.{Base, GraphIRReflection, ModuleInfo, IRContext}
+  import sigma.compiler.ir.core.MutableLazy
+  import sigma.compiler.ir.meta.ModuleInfo
+  import sigma.compiler.ir.{Base, GraphIRReflection, IRContext}
   import sigma.data.Nullable
   import sigma.reflection.RClass
